Implement photo manipulation with partners and authorization for partner controller methods

+++ EESTEC/Controllers/PartnerController.cs
@@ -1,6 +1,7 @@
 ﻿using EESTEC.Interfaces;
 using EESTEC.Models;
 using EESTEC.ViewModel;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 
@@ -10,11 +11,13 @@ namespace EESTEC.Controllers
     {
         private readonly IPartnerCategoryRepository _partnerCategoryRepository;
         private readonly IPartnerRepository _partnerRepository;
+        private readonly IPhotoService _photoService;
 
-        public PartnerController(IPartnerCategoryRepository partnerCategoryRepository, IPartnerRepository partnerRepository)
+        public PartnerController(IPartnerCategoryRepository partnerCategoryRepository, IPartnerRepository partnerRepository, IPhotoService photoService)
         {
             _partnerCategoryRepository = partnerCategoryRepository;
             _partnerRepository = partnerRepository;
+            _photoService = photoService;
         }
 
         private async Task<List<SelectListItem>> GetPartnerCategoriesAsync()
@@ -31,12 +34,15 @@ namespace EESTEC.Controllers
             }
             return list;
         }
+
+        [Authorize]
         public async Task<IActionResult> Index()
         {
             var partners = await _partnerRepository.GetAllAsync();
             return View(partners);
         }
 
+        [Authorize]
         public async Task<IActionResult> Create()
         {
             var partnerVM = new CreatePartnerViewModel
@@ -46,20 +52,22 @@ namespace EESTEC.Controllers
 
             return View(partnerVM);
         }
+        [Authorize]
         public async Task<IActionResult> Edit(int id)
         {
             var partner = await _partnerRepository.GetByIdAsync(id);
             if (partner == null)
                 return NotFound();
-            var partnerVM = new CreatePartnerViewModel
+            var partnerVM = new EditPartnerViewModel
             {
                 Name = partner.Name,
-                Image = partner.Image,
+                Image = null,
+                ImageUrl = partner.Image,
                 SelectedCategory = partner.PartnerCategory.Id.ToString(),
                 Website = partner.Website,
                 PartnerCategoriesSelectList = await GetPartnerCategoriesAsync(),
             };
-            foreach(var cat in partnerVM.PartnerCategoriesSelectList)
+            foreach (var cat in partnerVM.PartnerCategoriesSelectList)
             {
                 if (cat.Value == partnerVM.SelectedCategory)
                 {
@@ -70,6 +78,7 @@ namespace EESTEC.Controllers
             return View(partnerVM);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<IActionResult> Create(CreatePartnerViewModel partnerVM)
         {
@@ -78,6 +87,16 @@ namespace EESTEC.Controllers
                partnerVM.PartnerCategoriesSelectList = await GetPartnerCategoriesAsync();
                 return View(partnerVM);
             }
+
+            //Image
+            var result = await _photoService.AddPhotoAsync(partnerVM.Image);
+            if (result.Error != null)
+            {
+                ModelState.AddModelError("Photo", "Fotografija nije uspješno učitana");
+                partnerVM.PartnerCategoriesSelectList = await GetPartnerCategoriesAsync();
+                return View(partnerVM);
+            }
+
             var partnerCategory = await _partnerCategoryRepository.GetByIdAsync(Int32.Parse(partnerVM.SelectedCategory));
             if (partnerCategory == null)
             {
@@ -90,25 +109,29 @@ namespace EESTEC.Controllers
                 Name = partnerVM.Name,
                 PartnerCategory = partnerCategory,
                 Website = partnerVM.Website,
-                Image = partnerVM.Image,
+                Image = result.Url.ToString(),
             };
 
             _partnerRepository.Create(partner);
             return RedirectToAction("Index");
         }
 
+        [Authorize]
         [HttpPost]
-        public async Task<IActionResult> Edit(int id, CreatePartnerViewModel partnerVM)
+        public async Task<IActionResult> Edit(int id, EditPartnerViewModel partnerVM)
         {
+            var partner = await _partnerRepository.GetByIdAsync(id);
+            if (partner == null)
+                return NotFound();
+
             if (!ModelState.IsValid)
             {
                 ModelState.AddModelError("", "Greška pri uređivanju partnera!");
                 partnerVM.PartnerCategoriesSelectList = await GetPartnerCategoriesAsync();
+                partnerVM.ImageUrl = partner.Image;
                 return View("Edit", partnerVM);
             }
-            var partner = await _partnerRepository.GetByIdAsync(id);
-            if (partner == null)
-                return NotFound();
+            
 
             var partnerCategory = await _partnerCategoryRepository.GetByIdAsync(Int32.Parse(partnerVM.SelectedCategory));
             if (partnerCategory == null)
@@ -117,21 +140,42 @@ namespace EESTEC.Controllers
                 return View(partnerVM);
             }
 
+            if (partnerVM.Image != null)
+            {
+                await _photoService.DeletePhotoAsync(partner.Image);
+                var result = await _photoService.AddPhotoAsync(partnerVM.Image);
+                if (result.Error != null)
+                {
+                    ModelState.AddModelError("Photo", "Fotografija nije uspješno učitana");
+                    partnerVM.PartnerCategoriesSelectList = await GetPartnerCategoriesAsync();
+                    partnerVM.ImageUrl = partner.Image;
+                    return View(partnerVM);
+                }
+                partnerVM.ImageUrl = result.Url.ToString();
+
+            }
+
             partner.PartnerCategory = partnerCategory;
             partner.Website = partnerVM.Website;
-            partner.Image = partnerVM.Image;
+            partner.Image = partnerVM.ImageUrl;
             partner.Name = partnerVM.Name;
 
             _partnerRepository.Update(partner);
             return RedirectToAction("Index");
         }
 
+        [Authorize]
         [HttpPost, ActionName("Delete")]
         public async Task<IActionResult> Delete(int id)
         {
             var partner = await _partnerRepository.GetByIdAsync(id);
             if (partner == null)
                 return NotFound();
+            var result = await _photoService.DeletePhotoAsync(partner.Image);
+            if (result.Error != null)
+            {
+                return RedirectToAction("Index");
+            }
             _partnerRepository.Delete(partner);
             return RedirectToAction("Index");
         }
